+++ src/Testing/ZopeTestCase/ZopeLite.py
@@ -208,7 +208,6 @@ def _installPackage(name, quiet=0):
         else:
             if not quiet: _print('Installing %s ... NOT FOUND\n' % name)
 
-installProduct('PluginIndexes', 1)  # Must install first
 installProduct('OFSP', 1)
 
 # So people can use ZopeLite.app()